db moved to shared/lib, components moved to features and widgets
