+++ src/app/App.tsx
@@ -1,11 +1,11 @@
 import { useState } from 'react';
-import { useDatabase } from '../db/useDatabase';
+import Content from '../pages/Content/Content';
 import Login from '../pages/Login/Login';
 import Settings from '../pages/Settings/Settings';
 import { useAuth } from '../shared/hooks/useAuth';
+import { useDatabase } from '../shared/lib/db/useDatabase';
 import { WebsiteCard } from '../shared/types/types';
 import { ContentView } from '../shared/types/view';
-import Content from '../widgets/Content';
 import Sidebar from '../widgets/Sidebar';
 import Toolbar from '../widgets/Toolbar';
