+++ src/shared/lib/db/dbConfig.ts
