+++ src/shared/lib/db/useDatabase.ts
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
-import { uniqueId } from '../shared/lib/utils/generate';
-import { WebsiteCard, WebsiteCardCreate } from '../shared/types/types';
-import { DB_KEYS } from './keys';
+import { WebsiteCard, WebsiteCardCreate } from '../../types/types';
+import { uniqueId } from '../utils/generate';
+import { DB_KEYS } from './dbConfig';
 
 export const useDatabase = <T extends WebsiteCard>() => {
   const [state, setState] = useState<T[]>([]);
