+++ src/widgets/PasswordGeneratorModal.tsx
