+++ src/widgets/Sidebar.tsx
@@ -1,8 +1,8 @@
 import { FC, useState } from 'react';
-import PasswordGeneratorModal from '../components/PasswordGeneratorModal';
 import { useTranslations } from '../shared/hooks/useTranslations';
 import { ContentView } from '../shared/types/view';
 import Button from '../shared/ui/Button';
+import PasswordGeneratorModal from './PasswordGeneratorModal';
 
 interface Props {
   setView: (view: ContentView) => void;
